Extract waveform widget input
