+++ src/gui_app/waveform.rs
@@ -255,6 +255,7 @@ pub(super) use widget::WaveformWidgetProps;
 pub(super) use widget::{WaveformWidget, waveform_viewport_view};
 
 mod widget_geometry;
+mod widget_input;
 
 mod edit_fade_curve_paint;
 mod edit_fade_paint;
