+++ src/gui_app/waveform/widget_input.rs
@@ -1,0 +1,193 @@
+use radiant::{
+    gui::types::{Point, Rect},
+    widgets::{PointerButton, WidgetInput, WidgetOutput},
+};
+
+use super::{
+    WaveformActiveDragKind, WaveformEditFadeHandle, WaveformInteraction, WaveformSelectionKind,
+    WaveformWidget,
+};
+
+impl WaveformWidget {
+    pub(in crate::gui_app::waveform) fn handle_waveform_input(
+        &mut self,
+        bounds: Rect,
+        input: WidgetInput,
+    ) -> Option<WidgetOutput> {
+        match input {
+            WidgetInput::PointerMove { position } => {
+                self.common.state.hovered = bounds.contains(position);
+                self.active_drag_kind.map(|_| {
+                    WidgetOutput::typed(WaveformInteraction::UpdateSelection {
+                        visible_ratio: self.ratio_from_position(bounds, position),
+                    })
+                })
+            }
+            WidgetInput::Wheel { position, delta } if bounds.contains(position) => {
+                Some(WidgetOutput::typed(WaveformInteraction::Wheel {
+                    delta,
+                    anchor_ratio: self.ratio_from_position(bounds, position),
+                }))
+            }
+            WidgetInput::PointerPress {
+                position,
+                button: PointerButton::Primary,
+                ..
+            } if bounds.contains(position) => self.handle_primary_press(bounds, position),
+            WidgetInput::PointerDoubleClick {
+                position,
+                button: PointerButton::Primary,
+                ..
+            } if bounds.contains(position) => self.handle_primary_double_click(bounds, position),
+            WidgetInput::PointerPress {
+                position,
+                button: PointerButton::Secondary,
+                ..
+            } if bounds.contains(position) => self.handle_secondary_press(bounds, position),
+            WidgetInput::PointerPress {
+                position,
+                button: PointerButton::Auxiliary,
+                ..
+            } if bounds.contains(position) => {
+                Some(WidgetOutput::typed(WaveformInteraction::BeginPan {
+                    visible_ratio: self.ratio_from_position(bounds, position),
+                }))
+            }
+            WidgetInput::PointerRelease {
+                position,
+                button: PointerButton::Primary,
+                ..
+            } if self.primary_release_finishes_drag() => {
+                Some(WidgetOutput::typed(WaveformInteraction::FinishSelection {
+                    visible_ratio: self.ratio_from_position(bounds, position),
+                }))
+            }
+            WidgetInput::PointerRelease {
+                position,
+                button: PointerButton::Secondary,
+                ..
+            } if self.secondary_release_finishes_drag() => {
+                Some(WidgetOutput::typed(WaveformInteraction::FinishSelection {
+                    visible_ratio: self.ratio_from_position(bounds, position),
+                }))
+            }
+            WidgetInput::PointerRelease {
+                position,
+                button: PointerButton::Auxiliary,
+                ..
+            } if self.active_drag_kind == Some(WaveformActiveDragKind::Pan) => {
+                Some(WidgetOutput::typed(WaveformInteraction::FinishSelection {
+                    visible_ratio: self.ratio_from_position(bounds, position),
+                }))
+            }
+            _ => None,
+        }
+    }
+
+    fn ratio_from_position(&self, bounds: Rect, position: Point) -> f32 {
+        ((position.x - bounds.min.x) / bounds.width().max(1.0)).clamp(0.0, 1.0)
+    }
+
+    fn handle_primary_press(&self, bounds: Rect, position: Point) -> Option<WidgetOutput> {
+        if let Some(handle) = self.edit_fade_handle_at(bounds, position) {
+            return Some(WidgetOutput::typed(WaveformInteraction::BeginEditFade {
+                handle,
+                visible_ratio: self.ratio_from_position(bounds, position),
+            }));
+        }
+        if let Some(edge) =
+            self.selection_resize_handle_at(bounds, position, WaveformSelectionKind::Play)
+        {
+            return Some(WidgetOutput::typed(
+                WaveformInteraction::BeginSelectionResize {
+                    kind: WaveformSelectionKind::Play,
+                    edge,
+                    visible_ratio: self.ratio_from_position(bounds, position),
+                },
+            ));
+        }
+        if self.selection_move_handle_at(bounds, position, WaveformSelectionKind::Play) {
+            return Some(WidgetOutput::typed(
+                WaveformInteraction::BeginSelectionMove {
+                    kind: WaveformSelectionKind::Play,
+                    visible_ratio: self.ratio_from_position(bounds, position),
+                },
+            ));
+        }
+        if self.selection_move_handle_at(bounds, position, WaveformSelectionKind::Edit) {
+            return Some(WidgetOutput::typed(
+                WaveformInteraction::BeginSelectionMove {
+                    kind: WaveformSelectionKind::Edit,
+                    visible_ratio: self.ratio_from_position(bounds, position),
+                },
+            ));
+        }
+        Some(WidgetOutput::typed(WaveformInteraction::BeginSelection {
+            kind: WaveformSelectionKind::Play,
+            visible_ratio: self.ratio_from_position(bounds, position),
+        }))
+    }
+
+    fn handle_primary_double_click(&self, bounds: Rect, position: Point) -> Option<WidgetOutput> {
+        if let Some(
+            handle @ (WaveformEditFadeHandle::FadeInOuterStart
+            | WaveformEditFadeHandle::FadeOutOuterEnd),
+        ) = self.edit_fade_handle_at(bounds, position)
+        {
+            return Some(WidgetOutput::typed(
+                WaveformInteraction::ClearEditFadeSilence { handle },
+            ));
+        }
+        None
+    }
+
+    fn handle_secondary_press(&self, bounds: Rect, position: Point) -> Option<WidgetOutput> {
+        if let Some(handle) = self.edit_fade_handle_at(bounds, position) {
+            return Some(WidgetOutput::typed(WaveformInteraction::BeginEditFade {
+                handle,
+                visible_ratio: self.ratio_from_position(bounds, position),
+            }));
+        }
+        if self.selection_move_handle_at(bounds, position, WaveformSelectionKind::Edit) {
+            return Some(WidgetOutput::typed(
+                WaveformInteraction::BeginSelectionMove {
+                    kind: WaveformSelectionKind::Edit,
+                    visible_ratio: self.ratio_from_position(bounds, position),
+                },
+            ));
+        }
+        Some(WidgetOutput::typed(WaveformInteraction::BeginSelection {
+            kind: WaveformSelectionKind::Edit,
+            visible_ratio: self.ratio_from_position(bounds, position),
+        }))
+    }
+
+    fn primary_release_finishes_drag(&self) -> bool {
+        self.active_drag_kind
+            == Some(WaveformActiveDragKind::Selection(
+                WaveformSelectionKind::Play,
+            ))
+            || matches!(
+                self.active_drag_kind,
+                Some(
+                    WaveformActiveDragKind::EditFade(_)
+                        | WaveformActiveDragKind::SelectionResize(WaveformSelectionKind::Play, _)
+                        | WaveformActiveDragKind::SelectionMove(_)
+                )
+            )
+    }
+
+    fn secondary_release_finishes_drag(&self) -> bool {
+        self.active_drag_kind
+            == Some(WaveformActiveDragKind::Selection(
+                WaveformSelectionKind::Edit,
+            ))
+            || matches!(
+                self.active_drag_kind,
+                Some(
+                    WaveformActiveDragKind::EditFade(_)
+                        | WaveformActiveDragKind::SelectionMove(WaveformSelectionKind::Edit)
+                )
+            )
+    }
+}
